+++ site/src/index.jsx
@@ -25,7 +25,6 @@ footer:
       - { name: 'Buession Security', href: 'https://security.buession.com/' }
       - { name: 'Buession Logging', href: 'https://logging.buession.com/' }
       - { name: 'Buession SpringBoot', href: 'https://springboot.buession.com/' }
-      - { name: 'Buession SpringCloud', href: 'https://springcloud.buession.com/' }
       - { name: 'Buession Cas', href: 'https://cas.buession.com/' }
 
 ---