+++ CHANGELOG.md
@@ -9,6 +9,12 @@
 - [依赖库版本升级和安全漏洞修复](https://github.com/buession/buession-parent/releases/tag/v5.0.0)
 
 
+### 🔔 变化
+
+- DataSourceConfiguration 删除 GenericDataSource 初始化
+- 新增 DataSourceConsumer 在 DataSource bean 初始化时，可定制 DataSourceProperties
+
+
 ---